requirements.txt
torch
faiss-cpu
streamlit
langchain_groq
pypdf
numpy
langchain_google_genai
pillow
sentence-transformers
dotenv
transformers
langchain
scikit-learn
langchain_core
chromadb
langchain_community
pymupdf
google-generativeai
